refactor!: abstractgenerator -> abstract_generator
