+++ expertsystem/amplitude/__init__.py
@@ -1,11 +1,11 @@
 """All modules related to amplitude model creation."""
 
 __all__ = [
-    "abstractgenerator",
+    "abstract_generator",
     "canonical_decay",
     "helicity_decay",
 ]
 
-from . import abstractgenerator
+from . import abstract_generator
 from . import canonical_decay
 from . import helicity_decay
