+++ expertsystem/amplitude/abstract_generator.py
